# headless/headless-commerce/headless-commerce-admin-catalog-impl/rest-openapi.yaml
components:
...
    schemas:
...
        Error:
            properties:
                errorCode:
                    description:
                        Internal error code mapping
                    example: 996
                    readOnly: true
                    type: integer
                errorDescription:
                    example: Unable to find currency. Currency code should be expressed with
                        3-letter ISO 4217 format.
                    readOnly: true
                    type: string
                message:
                    example: No CommerceCurrency exists with the key {catalogId=41811, code=US
                        Dollar}
                    readOnly: true
                    type: string
                status:
                    description:
                        HTTP Status code
                    example: 404
                    readOnly: true
                    type: integer
            required:
                - errorCode
                - errorDescription
                - message
                - status
            type: object
...
        ProductConfiguration:
            properties:
                allowBackOrder:
                    example: true
                    type: boolean
                allowedOrderQuantities:
                    example:
                        - 10
                        - 20
                        - 30
                        - 40
                    items:
                        format: int32
                        type: integer
                    type: array
                displayAvailability:
                    example: true
                    type: boolean
                displayStockQuantity:
                    example: true
                    type: boolean
                inventoryEngine:
                    description:
                        The inventory engine that will be used to manage the product
                        inventory
                    type: string
                lowStockAction:
                    description:
                        The low stock action that will be performed when a product
                        is out of stock
                    type: string
                maxOrderQuantity:
                    format: int32
                    type: integer
                minOrderQuantity:
                    format: int32
                    type: integer
                minStockQuantity:
                    format: int32
                    type: integer
                multipleOrderQuantity:
                    format: int32
                    type: integer
            type: object
...
        ProductSubscriptionConfiguration:
            properties:
                enable:
                    example: true
                    type: boolean
                length:
                    example: 2
                    format: int32
                    type: integer
                numberOfLength:
                    example: 12
                    format: int64
                    type: integer
                subscriptionType:
                    enum:
                        - day
                        - month
                        - week
                        - year
                    example: month
                    type: string
                subscriptionTypeSettings:
                    additionalProperties:
                        type: string
                    example:
                        monthDay: "1"
                        monthlyMode: "0"
                    type: object
            type: object
...
    securitySchemes:
        BasicAuth:
            scheme: basic
            type: http
        BearerAuth:
            scheme: bearer
            type: http
        OAuth2:
            flows:
                authorizationCode:
                    authorizationUrl: /oauth/authorize
                    scopes:
                        HeadlessCommerceAdminCatalog.admin: Grants access to admin operations
                        HeadlessCommerceAdminCatalog.read: Grants read access
                        HeadlessCommerceAdminCatalog.write: Grants write access
                    tokenUrl: /oauth/token
            type: oauth2
info:
    contact:
        email: team-commerce@liferay.com
        name: Commerce Team
    description:         Liferay Commerce Admin Catalog API
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: Liferay Commerce Admin Catalog API
    version: v1.0
openapi: 3.0.0
paths:
...
    /products/:
        get:
            operationId: getProductsPage
            parameters:
                - in: query
                  name: filter
                  required: false
                  schema:
                      type: string
                - in: query
                  explode: true
                  name: page
                  required: false
                  schema:
                      type: integer
                  style: form
                - in: query
                  explode: true
                  name: pageSize
                  required: false
                  schema:
                      type: integer
                  style: form
                - in: query
                  name: sort
                  required: false
                  schema:
                      type: string
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Product'
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: '#/components/schemas/Product'
                                type: array
                    description:
                        Successful operation
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.read
            summary: Gets a list of Products.
            tags:
                - Product
        post:
            operationId: postProduct
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Product'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/Product'
                required: true
            responses:
                201:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Product'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Product'
                    description:
                        Created
                202:
                    description:
                        Accepted - Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Creates or updates a Product.
            tags:
                - Product
...
    /products/by-externalReferenceCode/{externalReferenceCode}/configuration/:
        get:
            operationId: getProductByExternalReferenceCodeConfiguration
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProductConfiguration'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/ProductConfiguration'
                    description:
                        Successful operation
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.read
            summary: Gets a Product Configuration of a Product.
            tags:
                - ProductConfiguration
        patch:
            operationId: patchProductByExternalReferenceCodeConfiguration
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ProductConfiguration'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/ProductConfiguration'
                required: true
            responses:
                202:
                    description:
                        Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Updates a Product Configuration of a Product.
            tags:
                - ProductConfiguration
...
    /products/by-externalReferenceCode/{externalReferenceCode}/subscriptionConfiguration/:
        get:
            operationId: getProductByExternalReferenceCodeSubscriptionConfiguration
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProductSubscriptionConfiguration'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/ProductSubscriptionConfiguration'
                    description:
                        Successful operation
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.read
            summary: Gets a Product Subscription Configuration of a Product.
            tags:
                - ProductSubscriptionConfiguration
        patch:
            operationId: patchProductByExternalReferenceCodeSubscriptionConfiguration
            parameters:
                - in: path
                  explode: false
                  name: externalReferenceCode
                  required: true
                  schema:
                      description:
                          The resource external reference code
                      example: AB-34098-789-N
                      minLength: 1
                      type: string
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ProductSubscriptionConfiguration'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/ProductSubscriptionConfiguration'
                required: true
            responses:
                202:
                    description:
                        Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Updates a Product Subscription Configuration of a Product.
            tags:
                - ProductSubscriptionConfiguration
...
    /products/{id}/configuration/:
        get:
            operationId: getProductIdConfiguration
            parameters:
                - in: path
                  explode: false
                  name: id
                  required: true
                  schema:
                      description:
                          The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
            responses:
                200:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ProductConfiguration'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/ProductConfiguration'
                    description:
                        Successful operation
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.read
            summary: Gets a Product by ID.
            tags:
                - ProductConfiguration
        patch:
            operationId: patchProductIdConfiguration
            parameters:
                - in: path
                  explode: false
                  name: id
                  required: true
                  schema:
                      description:
                          The resource Id
                      example: 30130
                      format: int64
                      minimum: 0
                      type: integer
                  style: simple
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ProductConfiguration'
                    application/xml:
                        schema:
                            $ref: '#/components/schemas/ProductConfiguration'
                required: true
            responses:
                202:
                    description:
                        Async
                204:
                    description:
                        Updated
                400:
                    description:
                        Invalid input
                401:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Authentication information is missing or invalid
                404:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        The specified resource was not found
                500:
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Error'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/Error'
                    description:
                        Unexpected error
            security:
                - BasicAuth: []
                - BearerAuth: []
                - OAuth2:
                      - HeadlessCommerceAdminCatalog.write
            summary: Updates a Product by ID.
            tags:
                - ProductConfiguration
...
servers:
    - description:
          SwaggerHub API Auto Mocking
      url: https://virtserver.swaggerhub.com/liferayinc/commerce-admin-catalog/v1.0